# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IK Part Number Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /segments/{key}:
    get:
      operationId: getSegment
      tags: [segments]
      summary: Get a single segment definition with its values
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Segment definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentDefinition"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /segments/{key}/values:
    post:
      operationId: addSegmentValue
      tags: [segments]
      summary: Add a new allowed value to a segment
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegmentValueInput"
      responses:
        "201":
          description: Value added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentValue"
        "409":
          description: Code already exists in segment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /segments/{key}/values/{code}:
    patch:
      operationId: updateSegmentValue
      tags: [segments]
      summary: Update description or applicability of a segment value
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegmentValueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentValue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSegmentValue
      tags: [segments]
      summary: Remove an allowed value from a segment
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SegmentValue:
      type: object
      required:
        - id
        - segmentKey
        - code
        - description
        - sortOrder
        - isActive
      properties:
        id:
          type: integer
        segmentKey:
          type: string
        code:
          type: string
        description:
          type: string
        applicableProducts:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
    SegmentDefinition:
      type: object
      required:
        - key
        - label
        - isRequired
        - isOptional
        - values
      properties:
        key:
          type: string
        label:
          type: string
        isRequired:
          type: boolean
        isOptional:
          type: boolean
        values:
          type: array
          items:
            $ref: "#/components/schemas/SegmentValue"
    SegmentValueInput:
      type: object
      required:
        - code
        - description
      properties:
        code:
          type: string
        description:
          type: string
        applicableProducts:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
    SegmentValueUpdate:
      type: object
      properties:
        description:
          type: string
        applicableProducts:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
...